refactor: move the element card to the components folder
